Favorite songs while you’re listening along
As a listener,
When I listen along with someone,
And I tap on the song I’m listening to at the bottom of the screen,
Then I see who is listening along with me,
And I see a heart icon.

When I tap the heart icon,
Then the song is added to my library,
And I a popup letting me know the song was added to my library.