Start Spotify if it’s not running
We can’t start playing music on any of the user’s Spotify-playing devices if none of them are currently playing music. The rule makes sense. Nobody should be able to reach into your phone at 4am and start playing music remotely.

This commit forces a user in listen along to tap a button which starts playing a silent track in Spotify if they’re not already listening to Spotify (giving us playback control without potentially subjecting the user to a genre they hate). It also informs them that they need to return to the Listen Along app.

The ability to start playing a song in Spotify  requires a redirect to Spotify and the user needs to know that they need to come back to Listen Along.