Don’t as persistently request Spotify be opened
Only tell people to start listening to music on Spotify when they try to listen to a broadcaster.

This is better is most circumstances; You should be able to view and scroll through all available broadcasters without being forced to listen to music on Spotify.

Currently you’re only prompted to open Spotify when you try to listen to a broadcaster by tapping on their name while you’re not listening to music on one of your Spotify devices.

The caveat to this being “good” most of the time is when you’re signing in for the first time. You’ll tap a broadcaster’s name, be redirected to the Spotify Sign In page, sign in, then be redirected back, and nothing will happen if you’re not listening to Spotify. If you are it will just work. Otherwise, you’ll have to tap the person’s name a second time, which isn’t optimal, to see the “Go Listen to Something in Spotify” popup.