Listen along immediately
As a listen along user,
Given I *am not signed in*,
Given I am not listening to Spotify,
When I tap the name “Nick Pachulski” on the Spotify users list screen,
And I am redirected to the Spotify sign in page,
When I enter my credentials,
And I tap “Sign In”,
Then Spotify begins playing the song that Nick Pachulski is listening to, scrubbed to the position in that song which Nick Pachulski is listening at,
So that I begin listening along immediately, without the need to be told to manually open Spotify, which is an extra, poor-user-experience step.

As a listen along user,
Given I *am signed in*,
Given I am not listening to Spotify,
When I tap the name “Nick Pachulski” on the Spotify users list screen,
And I am redirected to the Spotify sign in page,
When I enter my credentials,
And I tap “Sign In”,
Then Spotify begins playing the song that Nick Pachulski is listening to, scrubbed to the position in that song which Nick Pachulski is listening at,
So that I begin listening along immediately, without the need to be told to manually open Spotify, which is an extra, poor-user-experience step.