Disallow listening along with yourself
This is a very unique scenario, but…

Given I have signed in before,
Given I am listening to Spotify,
Given I am using a new device (which hasn’t been authenticated / associated with me yet),
When I open the Listen along app,
And I tap my own name in the list,
And I enter my Spotify credentials,
Then I am not set to be listening along with myself,
So that my name does not disappear from the list.

Your name would disappear from the list in this situation due to the implementation details of the app reacting to a state which should never have been allowed.

Listening along with yourself sets your user as listening along with someone else, which removes it from the list on your own device and would falsely indicate that you have 1 listener on other’s devices.

https://trello.com/c/C4bseYjE/25-you-can-listen-along-with-yourself-when-you-authenticate-on-a-new-device-disallow-this