Allow Spotify users to leave feedback
As a Spotify user using listen along,
When I open the app,
Then I see a support icon.

When I tap the support icon,
Then I see a feedback field with a “done” button.

When I tap “done”,
Then I see an “Are you sure you want to send this feedback?” Popup,
And the popup has two options: yes and no.

When I tap “yes”,
Then I am taken back to the Spotify users list screen,
And I see a popup thanking me for my feedback,

So that we as developers may receive input not apparent to ourselves.

https://trello.com/c/q7I4ov2h/45-add-feedback-form